+++ pwsh/dev/functions/addHtParameters.ps1
@@ -32,6 +32,7 @@ function addHtParameters {
         PolicyAtScopeOnly                            = [bool]$PolicyAtScopeOnly
         RBACAtScopeOnly                              = [bool]$RBACAtScopeOnly
         DoPSRule                                     = [bool]$DoPSRule
+        PSRuleFailedOnly                             = [bool]$PSRuleFailedOnly 
     }
     Write-Host 'htParameters:'
     $azAPICallConf['htParameters'] | format-table -AutoSize | Out-String